requirements.txt
Django==4.1.3
django-cors-headers==4.3.1
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
whitenoise==6.6.0
